refactor: use journeys admin query
