+++ apps/journeys-admin/src/components/JourneyList/ActiveJourneyList/ActiveJourneyList.tsx
@@ -10,7 +10,7 @@ import { useTranslation } from 'react-i18next'
 import { Dialog } from '@core/shared/ui/Dialog'
 
 import { JourneyStatus } from '../../../../__generated__/globalTypes'
-import { useAdminJourneysQuery } from '../../../libs/useAdminJourneysQuery'
+import { useAdminJourneysQuery } from '../../../libs/useJourneysAdminQuery'
 import { DiscoveryJourneys } from '../../DiscoveryJourneys'
 import { useTeam } from '../../Team/TeamProvider'
 import { JourneyCard } from '../JourneyCard'
