+++ apps/journeys-admin/src/components/JourneyList/JourneyList.stories.tsx
@@ -4,7 +4,7 @@ import { ComponentProps } from 'react'
 import { JourneyStatus } from '../../../__generated__/globalTypes'
 import { cache } from '../../libs/apolloClient/cache'
 import { journeysAdminConfig } from '../../libs/storybook'
-import { GET_ADMIN_JOURNEYS } from '../../libs/useAdminJourneysQuery/useAdminJourneysQuery'
+import { GET_ADMIN_JOURNEYS } from '../../libs/useJourneysAdminQuery/useJourneysAdminQuery'
 import { getDiscoveryJourneysMock } from '../DiscoveryJourneys/data'
 import { PageWrapper } from '../NewPageWrapper'
