+++ apps/journeys-admin/src/components/NewPageWrapper/NavigationDrawer/NavigationDrawer.stories.tsx
@@ -11,7 +11,7 @@ import {
   UserJourneyRole
 } from '../../../../__generated__/globalTypes'
 import { journeysAdminConfig } from '../../../libs/storybook'
-import { GET_ADMIN_JOURNEYS } from '../../../libs/useAdminJourneysQuery/useAdminJourneysQuery'
+import { GET_ADMIN_JOURNEYS } from '../../../libs/useJourneysAdminQuery/useJourneysAdminQuery'
 import { GET_USER_ROLE } from '../../../libs/useUserRoleQuery/useUserRoleQuery'
 
 import { GET_ME } from './NavigationDrawer'
