--- apps/journeys-admin/src/libs/useAdminJourneysQuery/index.ts
@@ -1,1 +1,0 @@
-export { useAdminJourneysQuery } from './useAdminJourneysQuery'
