+++ apps/journeys-admin/src/libs/useJourneysAdminQuery/index.ts
@@ -1,0 +1,1 @@
+export { useJourneysAdminQuery } from './useJourneysAdminQuery'
